+++ src/plugins/gs-plugin-packagekit-upgrade.c
@@ -71,7 +71,7 @@ gs_plugin_packagekit_progress_cb (PkProgress *progress,
 	} else if (type == PK_PROGRESS_TYPE_PERCENTAGE) {
 		gint percentage = pk_progress_get_percentage (progress);
 		if (percentage >= 0 && percentage <= 100)
-			gs_app_set_progress (data->app, percentage);
+			gs_app_set_progress (data->app, (guint) percentage);
 	}
 }
